+++ tests/AuthDigestTest.php
@@ -4,6 +4,7 @@ require_once("bootstrap.php");
 
 class AuthDigestTest extends PHPUnit_Framework_TestCase
 {
+
 	public function testEncode()
 	{
 		$cases = array(
@@ -16,6 +17,17 @@ class AuthDigestTest extends PHPUnit_Framework_TestCase
 		}
 	}
 
+	public function testVerifyCallback()
+	{
+		initKeys();
+		$mac1 = Qiniu_RequireMac(null);
+		$auth = 'QBox Vhiv6a22kVN_zhtetbPNeG9sY3JUL1HG597EmBwQ:JrRyg9So6DNrNDY5qj1sygt0SmQ=';
+		$url = 'http://rs.qbox.me/batch';
+		$body = 'op=/delete/cGhwc2RrOnRlc3RPcDI=&op=/delete/cGhwc2RrOnRlc3RPcDM=&op=/delete/cGhwc2RrOnRlc3RPcDQ=';
+		$pass = $mac1->VerifyCallback($auth, $url, $body);
+		$this->assertTrue($pass);
+	}
+	
 /*	public function testSetKeys()
 	{
 		$mac1 = Qiniu_RequireMac(null);